saneo/backend/requirements.txt
# Django
Django>=4.2,<5.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
django-cors-headers==4.3.1
django-filter==23.5

# Database
psycopg2-binary==2.9.9
dj-database-url==2.1.0

# Celery
celery==5.3.6
redis==5.0.1
django-celery-beat==2.5.0

# API Documentation
drf-spectacular==0.27.1

# Utilities
python-dotenv==1.0.1
Pillow==10.2.0

# Production
gunicorn==21.2.0
whitenoise==6.6.0

# Development
django-debug-toolbar==4.2.0
ipython==8.20.0
